fixed possible dataloss

+++ Source/Constructors/Trick.h
@@ -3,7 +3,7 @@
 	struct Trickconstructor
 	{
 		Trickconstructor() { }
-		Trickconstructor(uint64_t ID, uint32_t level, uint8_t apply) : TrickID(ID), Tricklvl(level), ApplyTrick(apply) { }
+		Trickconstructor(uint32_t ID, uint32_t level, uint8_t apply) : TrickID(ID), Tricklvl(level), ApplyTrick(apply) { }
 
 		uint32_t TrickID;
 		uint32_t Tricklvl;
